Move interfaces into own folder
