+++ src/Shared/Recruit.Vacancies.Client/Domain/Events/ApplicationReviewCreatedEvent.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using Esfa.Recruit.Vacancies.Client.Domain.Events.Interfaces;
 using Esfa.Recruit.Vacancies.Client.Domain.Messaging;
 using MediatR;
