+++ src/Shared/Recruit.Vacancies.Client/Domain/Events/Interfaces/IVacancyReviewEvent.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace Esfa.Recruit.Vacancies.Client.Domain.Events
+namespace Esfa.Recruit.Vacancies.Client.Domain.Events.Interfaces
 {
     public interface IVacancyReviewEvent
     {
