+++ src/Shared/Recruit.Vacancies.Client/Infrastructure/EventHandlers/UpdateDashboardOnVacancyChange.cs
@@ -4,6 +4,7 @@ using MediatR;
 using Microsoft.Extensions.Logging;
 using System.Threading;
 using System.Threading.Tasks;
+using Esfa.Recruit.Vacancies.Client.Domain.Events.Interfaces;
 using Esfa.Recruit.Vacancies.Client.Infrastructure.Services;
 
 namespace Esfa.Recruit.Vacancies.Client.Infrastructure.EventHandlers
